# keezenspel_openapi.yml
openapi: 3.1.0
info:
  title: Keezen Game API
  version: 1.0.0
  description: |
    OpenApi definition for the Keezen board game.
    This API manages players, games, moves, and cards.
servers:
  - url: /
    description: Same origin as the deployed app (local dev runs at http://localhost:4200)
paths:
...
  /trade/{sessionId}:
    post:
      summary: Team card trade — request, accept, reject or cancel a King/Ace swap
      operationId: teamTrade
      parameters:
        - $ref: '#/components/parameters/sessionId'
      tags:
        - trade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TradeAction'
      responses:
        '200':
          description: Trade action applied
        '400':
          description: Trade action was not allowed
        '404':
          description: Game not found
...
components:
  parameters:
    sessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    TradeAction:
      type: object
      description: |
        A team card-trade action from a player. REQUEST asks the teammate at once, carrying no
        card. OFFER (the requester) and ACCEPT (the teammate) then name each half of the swap, in
        whichever order they come — the trade completes as soon as both have landed. Either side
        may repeat their action to change their card while the other half is still missing.
        REJECT and CANCEL clear the trade outright.
      required:
        - action
        - playerId
      properties:
        action:
          type: string
          enum:
            - REQUEST
            - OFFER
            - ACCEPT
            - REJECT
            - CANCEL
        playerId:
          type: string
        card:
          $ref: '#/components/schemas/Card'
          description: The card given in return (OFFER) or the King/Ace handed over (ACCEPT)
...
    Card:
      type: object
      required:
        - suit
        - value
        - uuid
      properties:
        suit:
          description: |
            there is 1 suite per player
            with a maximum of 4 suites (0,1,2,3)
          type: integer
          minimum: 0
          maximum: 3
        value:
          description: the face value of the card 1 = ace, ... , 12 = king
          type: integer
        uuid:
          description: |
            a unique tracking number to distinguish between two identical cards (Ace♠︎),
            this happens when you have more than 4 players.
          type: integer
...